mlops-azure/drift-detection/requirements.txt
# Drift-Aware Retraining Pipeline - Python Dependencies

# Core
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Database
psycopg2-binary==2.9.9
pgvector==0.2.3

# OpenAI
openai==1.3.7

# Prometheus
prometheus-client==0.19.0

# ML/Data Science
matplotlib==3.7.2
seaborn==0.12.2

# Logging & Config
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Optional: EvidentlyAI for advanced drift detection
evidently==0.4.11
